WebserverConfig method refactoring return type Result
