+++ src/config/WebserverConfig.cpp
@@ -3,16 +3,21 @@
 WebserverConfig::WebserverConfig(FileDescriptor &file, char **envp) {
   err_meg = "";
   count_line = 0;
-  file_parsing(file, envp);
-  if (err_meg != "") {
-    if (!std::isdigit(static_cast<unsigned char>(err_meg[0])))
-      err_meg = ConfigError::add_line_number(count_line, err_meg);
+  Result<Void> result = file_parsing(file, envp);
+  if (!result.error().empty()) {
+    std::ostringstream oss;
+    oss << count_line;
+
+    err_meg = result.error();
+    if (!err_meg.empty() &&
+        !std::isdigit(static_cast<unsigned char>(err_meg[0])))
+      err_meg = oss.str() + " " + err_meg;
     return;
   }
   return;
 }
 
-bool WebserverConfig::file_parsing(FileDescriptor &file, char **envp) {
+Result<Void> WebserverConfig::file_parsing(FileDescriptor &file, char **envp) {
   std::string line;
   bool is_type_parse = false;
   bool is_cgi_parse = false;
@@ -24,7 +29,7 @@ bool WebserverConfig::file_parsing(FileDescriptor &file, char **envp) {
     count_line++;
     if (temp.error() != "") {
       err_meg = ConfigError::file_descriptor(temp.error());
-      return false;
+      return ERR(Void, ConfigError::file_descriptor(temp.error()));
     } else if (temp.value() == "")
       break;
     else if (temp.value() == "\n")
@@ -33,164 +38,142 @@ bool WebserverConfig::file_parsing(FileDescriptor &file, char **envp) {
     origin_line = utils::remove_char(temp.value(), '\n');
     err_meg = configutils::get_indent_whitespace_error(origin_line, 0);
     if (err_meg != "")
-      return false;
+      return ERR(Void, err_meg);
     line = utils::trim_whitespace(origin_line);
 
     if (line == "types =" || line == "types=") {
-      if (is_type_parse == true) {
-        err_meg = ConfigError::make(line, ERR_DUPLICATE_TYPE_BLOCK);
-        return false;
-      } else if (is_server_parse == true) {
-        err_meg = ConfigError::make(line, ERR_INVALID_GLOBAL_BLOCK_LOCATION);
-        return false;
-      } else if (!parse_types_block(file))
-        return false;
+      if (is_type_parse == true)
+        return ERR(Void, ConfigError::make(line, ERR_DUPLICATE_TYPE_BLOCK));
+      else if (is_server_parse == true)
+        return ERR(Void, ConfigError::make(line, ERR_INVALID_GLOBAL_BLOCK_LOCATION));
+      
+      Result<Void> result = parse_types_block(file);
+      if (!result.error().empty())
+        return result;
+      
       is_type_parse = true;
     } else if (WebserverConfig::is_server_config_header(line)) {
-      if (is_type_parse == false) {
-        err_meg = ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_TYPE_BLOCK_MISSING));
-        return false;
-      } else if (!parse_server_config_entry(file, line, envp))
-        return false;
+      if (is_type_parse == false)
+        return ERR(Void, ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_TYPE_BLOCK_MISSING)));
+      Result<Void> result = parse_server_config_entry(file, line, envp);
+      if (!result.error().empty())
+        return result;
+      
       is_server_parse = true;
     } else if (line == "cgi =" || line == "cgi=") {
-      if (is_cgi_parse == true) {
-        err_meg = ConfigError::make(origin_line, ERR_DUPLICATE_GLOBAL_CGI_BLOCK);
-        return false;
-      } else if (is_server_parse == true) {
-        err_meg = ConfigError::make(origin_line, ERR_INVALID_GLOBAL_BLOCK_LOCATION);
-        return false;
-      }
-      Result<Void> Result = RouteRule_CGI::parse_global_cgi_block(file, global_cgi,
+      if (is_cgi_parse == true)
+        return ERR(Void, ConfigError::make(origin_line, ERR_DUPLICATE_GLOBAL_CGI_BLOCK));
+      else if (is_server_parse == true)
+        return ERR(Void, ConfigError::make(origin_line, ERR_INVALID_GLOBAL_BLOCK_LOCATION));
+
+      Result<Void> result = RouteRule_CGI::parse_global_cgi_block(file, global_cgi,
                                                       count_line, envp);
-      if (Result.error() != "")
-        return false;
+      if (result.error() != "")
+        return result;
+      
       is_cgi_parse = true;
     } else if (line[0] == '!') {
-      if (is_err_page_parse == true) {
-        err_meg = ConfigError::make(origin_line, ERR_DUPLICATE_DEFAULT_ERROR_PAGE_BLOCK);
-        return false;
-      } else if (is_server_parse == true) {
-        err_meg = ConfigError::make(origin_line, ERR_INVALID_GLOBAL_BLOCK_LOCATION);
-        return false;
-      }
-      err_meg = ServerConfig::apply_err_page_entry(origin_line,
+      if (is_err_page_parse == true)
+        return ERR(Void, ConfigError::make(origin_line, ERR_DUPLICATE_DEFAULT_ERROR_PAGE_BLOCK));
+      else if (is_server_parse == true)
+        return ERR(Void, ConfigError::make(origin_line, ERR_INVALID_GLOBAL_BLOCK_LOCATION));
+
+      Result<Void> result = ServerConfig::apply_err_page_entry(origin_line,
           line, default_err_page, envp);
-      if (err_meg != "") 
-        return false;
+      if (result.error() != "")
+        return result;
+      
       is_err_page_parse = true;
-    } else {
-      err_meg = ConfigError::make(origin_line, ERR_INVALID_TOP_LEVEL_FORMAT);
-      return false;
-    }
+    } else
+      return ERR(Void, ConfigError::make(origin_line, ERR_INVALID_TOP_LEVEL_FORMAT));
   }
 
-  if (type_map.empty()) {
-    err_meg = ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_TYPE_BLOCK_MISSING));
-    return false;
-  } else if (serverconfig_map.empty()) {
-    err_meg = ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_SERVER_BLOCK_MISSING));
-    return false;
-  }
-  return true;
+  if (type_map.empty())
+    return ERR(Void, ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_TYPE_BLOCK_MISSING)));
+  else if (serverconfig_map.empty())
+    return ERR(Void, ConfigError::add_line_number(1, ConfigError::make("", "", ERR_REQUIRED_SERVER_BLOCK_MISSING)));
+  return OKV;
 }
 
-std::vector<std::string>
+Result<std::vector<std::string> >
 WebserverConfig::parse_type_keys(const std::string &key) {
   int number_of_key = 0;
   std::string temp = key;
   std::vector<std::string> key_data;
 
-  if (utils::has_invalid_char(temp, "_|")) {
-    err_meg = ConfigError::make(origin_line, temp, ERR_INVALID_MIME_EXTENSION_CHAR);
-    return (key_data);
-  }
+  if (utils::has_invalid_char(temp, "_|"))
+    return ERR(std::vector<std::string>, ConfigError::make(origin_line, temp, ERR_INVALID_MIME_EXTENSION_CHAR));
 
   key_data = utils::string_split(temp, "|");
   number_of_key = utils::count_occurrences(temp, "|") + 1;
-  if (key_data.size() != static_cast<std::size_t>(number_of_key)) {
-    err_meg = ConfigError::make(origin_line, temp, ERR_MIME_EXTENSION_COUNT_MISMATCH);
-    return (std::vector<std::string>());
-  }
+  if (key_data.size() != static_cast<std::size_t>(number_of_key))
+    ERR(std::vector<std::string>, ConfigError::make(origin_line, temp, ERR_MIME_EXTENSION_COUNT_MISMATCH));
   for (std::size_t i = 0; i < key_data.size(); ++i) {
-    if (type_map.find(key_data[i]) != type_map.end()) {
-      err_meg = ConfigError::make(origin_line, key_data[i], ERR_DUPLICATE_MIME_EXTENSION);
-      return (std::vector<std::string>());
-    }
+    if (type_map.find(key_data[i]) != type_map.end())
+      ERR(std::vector<std::string>, ConfigError::make(origin_line, key_data[i], ERR_DUPLICATE_MIME_EXTENSION));
   }
-  return (key_data);
+  return OK(std::vector<std::string>, key_data);
 }
 
-bool WebserverConfig::is_valid_mime_type(const std::string &value) {
+Result<Void> WebserverConfig::is_valid_mime_type(const std::string &value) {
   std::vector<std::string> value_data;
 
-  if (utils::has_invalid_char(value, "/-")) {
-    err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_CHAR);
-    return false;
-  } else if (value[0] == '-' || value[0] == '/') {
-    err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-    return false;
-  }
-  for (std::size_t i = 1; i < value.size(); ++i) {
+  if (utils::has_invalid_char(value, "/-"))
+    return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_CHAR));
+  else if (value[0] == '-' || value[0] == '/')
+    return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
+  
+  for (std::size_t i = 1; i < value.size(); ++i)
     if (value[i] == '-' && value[i - 1] == '-') {
-      err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-      return false;
-    }
+      return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
   }
+
   value_data = utils::string_split(value, "/");
-  if (value_data.size() != 2) {
-    err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-    return false;
-  } else if (utils::count_occurrences(value, "/") != 1) {
-    err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-    return false;
-  }
+  if (value_data.size() != 2)
+    return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
+  else if (utils::count_occurrences(value, "/") != 1)
+    return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
 
   for (std::size_t i = 0; i < value.size(); ++i) {
     if (value[i] == '-') {
-      if (i == value.size() - 1) {
-        err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-        return false;
-      }
+      if (i == value.size() - 1)
+        return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
+      
       if (!std::isalnum(static_cast<unsigned char>(value[i - 1])) ||
-          !std::isalnum(static_cast<unsigned char>(value[i + 1]))) {
-        err_meg = ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT);
-        return false;
-      }
+          !std::isalnum(static_cast<unsigned char>(value[i + 1])))
+        return ERR(Void, ConfigError::make(origin_line, value, ERR_INVALID_MIME_TYPE_FORMAT));
     }
   }
-  return true;
+  return OKV;
 }
 
-bool WebserverConfig::parse_type_mapping(const std::string &line,
+Result<Void> WebserverConfig::parse_type_mapping(const std::string &line,
                                          std::vector<std::string> &keys_out,
                                          std::string &value_out) {
   std::size_t pos = line.find("->");
-  if (pos == std::string::npos) {
-    err_meg = ConfigError::make(origin_line, line, ERR_MISSING_MIME_MAPPING_OPERATOR);
-    return false;
-  } else if (utils::count_occurrences(line, "->") != 1) {
-    err_meg = ConfigError::make(origin_line, line, ERR_INVALID_MIME_MAPPING_SYNTAX);
-    return false;
-  }
+  if (pos == std::string::npos)
+    return ERR(Void, ConfigError::make(origin_line, line, ERR_MISSING_MIME_MAPPING_OPERATOR));
+  else if (utils::count_occurrences(line, "->") != 1)
+    return ERR(Void, ConfigError::make(origin_line, line, ERR_INVALID_MIME_MAPPING_SYNTAX));
+  
   std::vector<std::string> type_data = utils::string_split(line, "->");
-  if (type_data.size() != 2) {
-    err_meg = ConfigError::make(origin_line, line, ERR_INVALID_MIME_MAPPING_VALUE);
-    return false;
-  }
+  if (type_data.size() != 2)
+    return ERR(Void, ConfigError::make(origin_line, line, ERR_INVALID_MIME_MAPPING_VALUE));
 
-  std::vector<std::string> keys =
-      WebserverConfig::parse_type_keys(utils::trim_whitespace(type_data[0]));
-  if (keys.empty() || !is_valid_mime_type(utils::trim_whitespace(type_data[1])))
-    return false;
-  err_meg = "";
+  Result<std::vector<std::string> > keys_result = parse_type_keys(utils::trim_whitespace(type_data[0]));
+  if (!keys_result.error().empty())
+    return ERR(Void, keys_result.error());
+  
+    Result<Void> mime_result = is_valid_mime_type(utils::trim_whitespace(type_data[1]));
+  if (!mime_result.error().empty())
+    return mime_result;
 
-  keys_out = keys;
+  keys_out = keys_result.value();
   value_out = utils::trim_whitespace(type_data[1]);
-  return true;
+  return OKV;
 }
 
-bool WebserverConfig::parse_types_block(FileDescriptor &file) {
+Result<Void> WebserverConfig::parse_types_block(FileDescriptor &file) {
   std::string line;
   std::string value;
   std::vector<std::string> keys;
@@ -200,40 +183,35 @@ bool WebserverConfig::parse_types_block(FileDescriptor &file) {
     count_line++;
     if (temp.error() != "") {
       err_meg = ConfigError::file_descriptor(temp.error());
-      return (false);
+      return ERR(Void, ConfigError::file_descriptor(temp.error()));
     } else if (temp.value() == "\n" || temp.value() == "")
       break;
 
     origin_line = utils::remove_char(temp.value(), '\n');
-    err_meg = configutils::get_indent_whitespace_error(origin_line, 1);
-    if (err_meg != "")
-      return false;
+    std::string err = configutils::get_indent_whitespace_error(origin_line, 1);
+    if (err != "")
+      return ERR(Void, err);
     line = utils::trim_whitespace(origin_line);
 
-    if (!parse_type_mapping(line, keys, value))
-      return false;
+    Result<Void> mapping_result = parse_type_mapping(line, keys, value);
+    if (!mapping_result.error().empty())
+      return mapping_result;
 
     for (std::size_t i = 0; i < keys.size(); ++i) {
       const std::string &k = keys[i];
       if (k == "_") {
-        if (!default_mime.empty()) {
-          err_meg = ConfigError::make(origin_line, value, ERR_DUPLICATE_DEFAULT_MIME_TYPE);
-          return false;
-        }
+        if (!default_mime.empty())
+          return ERR(Void, ConfigError::make(origin_line, value, ERR_DUPLICATE_DEFAULT_MIME_TYPE));
         default_mime = value;
         continue;
-      } else if (type_map.find(k) != type_map.end()) {
-        err_meg = ConfigError::make(origin_line, k, ERR_DUPLICATE_MIME_EXTENSION);
-        return false;
-      }
+      } else if (type_map.find(k) != type_map.end()) 
+        return ERR(Void, ConfigError::make(origin_line, k, ERR_DUPLICATE_MIME_EXTENSION));
       type_map[k] = value;
     }
   }
-  if (default_mime.empty()) {
-    err_meg = ConfigError::make("", "", ERR_MISSING_DEFAULT_MIME_TYPE);
-    return false;
-  }
-  return true;
+  if (default_mime.empty())
+    return ERR(Void, ConfigError::make("", "", ERR_MISSING_DEFAULT_MIME_TYPE));
+  return OKV;
 }
 
 bool WebserverConfig::is_server_config_header(const std::string &line) {
@@ -258,7 +236,7 @@ bool WebserverConfig::is_server_config_header(const std::string &line) {
   return (i == line.size());
 }
 
-bool WebserverConfig::parse_server_config_entry(FileDescriptor &file,
+Result<Void> WebserverConfig::parse_server_config_entry(FileDescriptor &file,
                                                 const std::string &line,
                                                 char **envp) {
   unsigned int key = 0;
@@ -266,33 +244,25 @@ bool WebserverConfig::parse_server_config_entry(FileDescriptor &file,
   ServerConfig server(file, global_cgi, envp);
 
   err_meg = configutils::string_to_unsigned_int(WebserverConfig::parse_server_port(temp), key);
-  if (err_meg != "") {
-    err_meg = ConfigError::make(line, WebserverConfig::parse_server_port(temp), err_meg);
-    return false;
-  }
-  if (MIN_PORT_VALUE > key || key > MAX_PORT_VALUE) {
-    err_meg = ConfigError::make(line, WebserverConfig::parse_server_port(temp), ERR_INVALID_SERVER_PORT);
-    return false;
-  }
+  if (err_meg != "")
+    return ERR(Void, ConfigError::make(line, WebserverConfig::parse_server_port(temp), err_meg));
+  if (MIN_PORT_VALUE > key || key > MAX_PORT_VALUE)
+    return ERR(Void, ConfigError::make(line, WebserverConfig::parse_server_port(temp), ERR_INVALID_SERVER_PORT));
 
-  if (server.get_err_meg() != "") {
-    err_meg = server.get_err_meg();
-    count_line += server.get_count_line();
-    return false;
-  } else if (server.get_routes().size() == 0 &&
+  if (server.get_err_meg() != "")
+    return ERR(Void, server.get_err_meg());
+  else if (server.get_routes().size() == 0 &&
              server.get_route_rule_cgi().size() == 0) {
-    err_meg = ConfigError::make(origin_line, ERR_EMPTY_SERVER_BLOCK);
-    return false;
+    return ERR(Void, ConfigError::make(origin_line, ERR_EMPTY_SERVER_BLOCK));
   } else if (serverconfig_map.find(key) != serverconfig_map.end()) {
     std::ostringstream oss;
     oss << key;
 
-    err_meg = ConfigError::make(line, oss.str(), ERR_DUPLICATE_SERVER_PORT);
-    return false;
+    return ERR(Void, ConfigError::make(line, oss.str(), ERR_DUPLICATE_SERVER_PORT));
   }
   count_line += server.get_count_line();
   serverconfig_map[key] = server;
-  return true;
+  return OKV;
 }
 
 std::string WebserverConfig::parse_server_port(const std::string &key) {
